requirements.txt
beautifulsoup4==4.10.0
discord.py==1.7.3
playwright==v1.17.2
